feat: Negative infinity and math engine double patch

+++ PhoenixLang/Math/MathEngine.cs
@@ -22,7 +22,7 @@ public static class MathEngine
         }
         catch (InvalidCastException e)
         {
-            return int.TryParse(expr.Evaluate().ToString(), out var final) ? final : default(double);
+            return double.TryParse(expr.Evaluate().ToString(), out var final) ? final : default(double);
         }
     }   
 }
